Split Integration Tests in multiple files
